requirements.txt
pycryptodome
imageio
imageio-ffmpeg
numpy